+++ Modules/Intent.Modules.OpenTelemetry/Templates/TemplateExtensions.cs
@@ -1,0 +1,20 @@
+using System.Collections.Generic;
+using Intent.Modules.Common.Templates;
+using Intent.Modules.OpenTelemetry.Templates.OpenTelemetryConfiguration;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: DefaultIntentManaged(Mode.Fully, Targets = Targets.Usings)]
+[assembly: IntentTemplate("Intent.ModuleBuilder.Templates.TemplateExtensions", Version = "1.0")]
+
+namespace Intent.Modules.OpenTelemetry.Templates
+{
+    public static class TemplateExtensions
+    {
+        public static string GetOpenTelemetryConfigurationName<T>(this IntentTemplateBase<T> template)
+        {
+            return template.GetTypeName(OpenTelemetryConfigurationTemplate.TemplateId);
+        }
+
+    }
+}