+++ Tests/Publish.AspNetCore.GooglePubSub.TestApplication/Publish.AspNetCore.GooglePubSub.TestApplication.Application/Common/Eventing/IEventBus.cs
@@ -9,8 +9,8 @@ namespace Publish.AspNetCore.GooglePubSub.TestApplication.Application.Common.Eve
 {
     public interface IEventBus
     {
-        void Publish<T>(T message) where T : class;
 
+        void Publish<T>(T message) where T : class;
         Task FlushAllAsync(CancellationToken cancellationToken = default);
     }
 }