+++ Tests/Publish.AspNetCore.GooglePubSub.TestApplication/Publish.AspNetCore.GooglePubSub.TestApplication.Application/Common/Eventing/IIntegrationEventHandler.cs
@@ -11,6 +11,7 @@ namespace Publish.AspNetCore.GooglePubSub.TestApplication.Application.Common.Eve
     public interface IIntegrationEventHandler<TMessage>
         where TMessage : class
     {
+
         Task HandleAsync(TMessage message, CancellationToken cancellationToken = default);
     }
 }