+++ Tests/Publish.AspNetCore.MassTransit.OutBoxNone.TestApplication/Publish.AspNetCore.MassTransit.OutBoxNone.Application/Common/Eventing/IIntegrationEventHandler.cs
@@ -11,6 +11,7 @@ namespace Publish.AspNetCore.MassTransit.OutBoxNone.Application.Common.Eventing
     public interface IIntegrationEventHandler<TMessage>
         where TMessage : class
     {
+
         Task HandleAsync(TMessage message, CancellationToken cancellationToken = default);
     }
 }