+++ Tests/Subscribe.GooglePubSub.TestApplication/Subscribe.GooglePubSub.TestApplication.Application/Common/Eventing/IEventBus.cs
@@ -9,8 +9,8 @@ namespace Subscribe.GooglePubSub.TestApplication.Application.Common.Eventing
 {
     public interface IEventBus
     {
-        void Publish<T>(T message) where T : class;
 
+        void Publish<T>(T message) where T : class;
         Task FlushAllAsync(CancellationToken cancellationToken = default);
     }
 }