+++ Tests/Subscribe.MassTransit.TestApplication/Subscribe.MassTransit.TestApplication.Application/Common/Eventing/IIntegrationEventHandler.cs
@@ -11,6 +11,7 @@ namespace Subscribe.MassTransit.TestApplication.Application.Common.Eventing
     public interface IIntegrationEventHandler<TMessage>
         where TMessage : class
     {
+
         Task HandleAsync(TMessage message, CancellationToken cancellationToken = default);
     }
 }